automation: Added AutomationHandler class.
This class now only contains dummy code for reading and writting settings to a json file. Additionally, automation.c has been changed to a c++ class.

For now, all signals and the separate thread are created in this file as I could not figure out how to do this from the
AutomationHandler class.